moved logic to MainView, cards are automatically drawn on turn change.
