+++ client/components/Controller/TreatView.js
@@ -1,6 +1,6 @@
 /* eslint-disable react/button-has-type */
 import React from 'react'
-import {treatInfection, updateActions} from '../../funcs/utils'
+import {treatInfection} from '../../funcs/utils' //, updateActions
 import db from '../../../server/db'
 import CURRENT_GAME from '../../../secrets'
 
@@ -34,7 +34,7 @@ export const TreatView = props => {
           )
           // decrementActions(props.playerId, props.playerInfo.actions)
           // props.nextTurn(props.playerInfo, `${props.playerId}`)
-          updateActions(props.playerId)
+          props.updateActions(props.playerId)
         }}
       >
         TREAT
@@ -56,7 +56,7 @@ export const TreatView = props => {
           )
           // decrementActions(props.playerId, props.playerInfo.actions)
           // props.nextTurn(props.playerInfo, `${props.playerId}`)
-          updateActions(props.playerId)
+          props.updateActions(props.playerId)
         }}
       >
         TREAT
@@ -78,7 +78,7 @@ export const TreatView = props => {
           )
           // decrementActions(props.playerId, props.playerInfo.actions)
           // props.nextTurn(props.playerInfo, `${props.playerId}`)
-          updateActions(props.playerId)
+          props.updateActions(props.playerId)
         }}
       >
         TREAT
@@ -100,7 +100,7 @@ export const TreatView = props => {
           )
           // decrementActions(props.playerId, props.playerInfo.actions)
           // props.nextTurn(props.playerInfo, `${props.playerId}`)
-          updateActions(props.playerId)
+          props.updateActions(props.playerId)
         }}
       >
         TREAT
