+++ client/funcs/utils.js
@@ -3,6 +3,12 @@
 const infectionDeck = require('../data/infectionDeck')
 const db = require('../../server/db')
 const CURRENT_GAME = require('../../secrets')
+// import MainView from '../components/Controller/MainView'
+
+// const Player1MainView = new MainView()
+// const Player2MainView = new MainView()
+// const Player3MainView = new MainView()
+// const Player4MainView = new MainView()
 
 const game = db.collection('rooms').doc(CURRENT_GAME)
 
@@ -263,54 +269,53 @@ const researchCure = async (playerId, hand, playerDiscard, infectionStatus) => {
   }
 }
 
-const updateActions = async player => {
-  const docRef = await game.get()
-  const data = await docRef.data()
+// const updateActions = async player => {
+//   const docRef = await game.get()
+//   const data = await docRef.data()
 
-  const whoIsNext = {
-    player1Info: 'player2Info',
-    player2Info: 'player3Info',
-    player3Info: 'player4Info',
-    player4Info: 'player1Info'
-  }
+//   const whoIsNext = {
+//     player1Info: 'player2Info',
+//     player2Info: 'player3Info',
+//     player3Info: 'player4Info',
+//     player4Info: 'player1Info'
+//   }
 
-  const currentPlayerInfo = data[player]
-  const nextPlayer = whoIsNext[player]
-  const nextPlayerInfo = data[nextPlayer]
+//   const currentPlayerInfo = data[player]
+//   const nextPlayer = whoIsNext[player]
+//   const nextPlayerInfo = data[nextPlayer]
 
-  if (currentPlayerInfo.actions > 1 && currentPlayerInfo.isTurn === true) {
-    currentPlayerInfo.actions--
-    console.log('actions > 1, decrement only', player, currentPlayerInfo)
-    await game.set(
-      {
-        [player]: currentPlayerInfo
-      },
-      {merge: true}
-    )
-  } else if (
-    currentPlayerInfo.actions <= 1 &&
-    currentPlayerInfo.isTurn === true
-  ) {
-    currentPlayerInfo.actions = 0
-    currentPlayerInfo.isTurn = false
-    nextPlayerInfo.actions = 4
-    nextPlayerInfo.isTurn = true
-    console.log(
-      'actions = 1, set to 0 and switch players',
-      player,
-      currentPlayerInfo,
-      nextPlayer,
-      nextPlayerInfo
-    )
-    await game.set(
-      {
-        [player]: currentPlayerInfo,
-        [nextPlayer]: nextPlayerInfo
-      },
-      {merge: true}
-    )
-  }
-}
+//   if (currentPlayerInfo.actions > 1 && currentPlayerInfo.isTurn === true) {
+//     currentPlayerInfo.actions--
+//     console.log('actions > 1, decrement only', player, currentPlayerInfo)
+//     await game.set(
+//       {
+//         [player]: currentPlayerInfo
+//       },
+//       { merge: true }
+//     )
+//   } else if (
+//     currentPlayerInfo.actions <= 1 &&
+//     currentPlayerInfo.isTurn === true
+//   ) {
+//     currentPlayerInfo.actions = 0
+//     currentPlayerInfo.isTurn = false
+//     nextPlayerInfo.actions = 4
+//     nextPlayerInfo.isTurn = true
+
+//     await game.set(
+//       {
+//         [player]: currentPlayerInfo,
+//         [nextPlayer]: nextPlayerInfo
+//       },
+//       { merge: true }
+//     )
+
+//     await this.drawInfectionCard()
+//     await this.drawInfectionCard()
+//     await this.drawPlayerCard()
+//     await this.drawPlayerCard()
+//   }
+// }
 
 module.exports = {
   shuffle,
@@ -321,6 +326,6 @@ module.exports = {
   discardPlayerCard,
   addEpidemics,
   researchCure,
-  resetDidOutbreak,
-  updateActions
+  resetDidOutbreak
+  // updateActions
 }
